Moving all language/fluent syntax flow types into a Flow namespace
Note that this doesn't affect the end-user in 99% of scenarios, since this is a fluent syntax. Meaning, they don't directly reference these flow types by name (which would require importing the namespace). Since we chain them together in calls, we don't need to know their name or namespace.
