+++ src/NanoBuilder/NanoBuilder/IParameterComposer.cs
@@ -1,3 +1,5 @@
+using NanoBuilder.Flow;
+
 namespace NanoBuilder
 {
    /// <summary>
