+++ src/NanoBuilder/NanoBuilder/ITypeMapper.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using NanoBuilder.Flow;
 
 namespace NanoBuilder
 {
